requirements.txt
langchain-anthropic
langgraph
google-analytics-data
python-dotenv
pandas
fpdf2